% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_meters.R
\name{in_meters}
\alias{in_meters}
\title{Mark step values as meters}
\usage{
in_meters(x)
}
\arguments{
\item{x}{A single positive numeric value.}
}
\value{
A \link[units:units]{units::units} object in meters (\code{m}).
}
\description{
Creates a units value in meters for use in \code{step} arguments.
Use plain numerics for cell units, and \code{in_meters()} for map-distance steps.
}
\examples{
in_meters(100)

}
