% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_plot_iter_diagnostics.R
\name{sc_plot_iter_diagnostics}
\alias{sc_plot_iter_diagnostics}
\title{Plot iteration diagnostics}
\usage{
sc_plot_iter_diagnostics(x)
}
\arguments{
\item{x}{A supercells object with an \code{iter_diagnostics} attribute,
or a diagnostics list containing \code{mean_distance}}
}
\value{
Invisibly returns \code{TRUE} when a plot is created
}
\description{
Plot mean distance across iterations for a supercells run
}
\examples{
library(supercells)
vol = terra::rast(system.file("raster/volcano.tif", package = "supercells"))
vol_sc = sc_slic_points(vol, step = 8, compactness = 1, iter_diagnostics = TRUE)
sc_plot_iter_diagnostics(vol_sc)
}
\seealso{
\code{\link[=sc_slic]{sc_slic()}}, \code{\link[=sc_slic_points]{sc_slic_points()}}
}
