% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_slic_params.R
\name{sc_slic_set_params}
\alias{sc_slic_set_params}
\title{Set stored \code{sc_slic()} parameters}
\usage{
sc_slic_set_params(sc, params)
}
\arguments{
\item{sc}{An sf object.}

\item{params}{A data.frame, typically from \code{\link[=sc_slic_get_params]{sc_slic_get_params()}}. Only the
first row is used.}
}
\value{
The input object with updated attributes.
}
\description{
Writes key \code{sc_slic()} parameters to attributes on a supercells object.
}
\seealso{
\code{\link[=sc_slic]{sc_slic()}}, \code{\link[=sc_slic_get_params]{sc_slic_get_params()}}
}
