% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercell.R
\name{supercell}
\alias{supercell}
\title{Title}
\usage{
supercell(x, step, nc, dist_fun = "euclidean", clean = TRUE)
}
\arguments{
\item{x}{a}

\item{step}{a}

\item{nc}{a}

\item{dist_fun}{a}

\item{clean}{a}
}
\value{
a
}
\description{
Title
}
\examples{
#a
}
