% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercells.R
\name{supercells}
\alias{supercells}
\title{Creates supercells}
\usage{
supercells(
  x,
  k,
  compactness,
  dist_fun = "euclidean",
  avg_fun = "mean",
  clean = TRUE,
  iter = 10,
  transform = NULL,
  step,
  minarea,
  metadata = TRUE,
  chunks = FALSE,
  verbose = 0
)
}
\arguments{
\item{x}{An object of class SpatRaster (terra) or class stars (stars)}

\item{k}{The number of supercells desired by the user (the output number can be slightly different!).
You can use either \code{k} or \code{step}.
It is also possible to provide a set of points (an \code{sf} object) as \code{k} together with the \code{step} value to create custom cluster centers.}

\item{compactness}{A compactness value. Larger values cause supercells to be more compact/even (square).
Use \code{\link[=use_adaptive]{use_adaptive()}} to enable adaptive compactness (ASLIC).
A compactness value depends on the range of input cell values and selected distance measure.}

\item{dist_fun}{A distance function. Currently implemented distance functions are \code{"euclidean"}, \code{"jsd"}, \code{"dtw"} (dynamic time warping), name of any distance function from the \code{philentropy} package (see \code{\link[philentropy:getDistMethods]{philentropy::getDistMethods()}}; "log2" is used in this case), or any user defined function accepting two vectors and returning one value. Default: \code{"euclidean"}}

\item{avg_fun}{An averaging function - how the values of the supercells' centers are calculated? The algorithm internally implements common functions \code{"mean"} and \code{"median"} (provided with quotation marks), but also accepts any fitting R function (e.g., \code{base::mean()} or \code{stats::median()}, provided as plain function name: \code{mean}). Default: \code{"mean"}. See details for more information.}

\item{clean}{Should connectivity of the supercells be enforced?}

\item{iter}{The number of iterations performed to create the output.}

\item{transform}{Transformation to be performed on the input. By default, no transformation is performed. Currently available transformation is "to_LAB": first, the conversion from RGB to the LAB color space is applied, then the supercells algorithm is run, and afterward, a reverse transformation is performed on the obtained results. (This argument is experimental and may be removed in the future).}

\item{step}{Initial center spacing. You can use either \code{k} or \code{step}.
Provide a plain numeric value for cell units, or use \code{\link[=use_meters]{use_meters()}} for
map-distance steps in meters (automatically converted to cells using raster resolution).}

\item{minarea}{Specifies the minimal size of a supercell (in cells). Only works when \code{clean = TRUE}.
By default, when \code{clean = TRUE}, average area (A) is calculated based on the total number of cells divided by a number of supercells
Next, the minimal size of a supercell equals to A/(2^2) (A is being right shifted)}

\item{metadata}{Logical. Controls whether metadata columns
("supercells", "x", "y") are included.}

\item{chunks}{Should the input (\code{x}) be processed with chunking before deriving supercells?
Use \code{FALSE} (default), \code{TRUE} (automatic chunking based on size), or a numeric value
(chunk side length in number of cells). When \code{TRUE}, the memory limit can be set
with \code{options(supercells.chunk_mem_gb = 4)}.}

\item{verbose}{An integer specifying the level of text messages printed during calculations. 0 means no messages (default), 1 provides basic messages (e.g., calculation stage).}
}
\value{
An sf object with several columns: (1) supercells - an id of each supercell, (2) y and x coordinates, (3) one or more columns with average values of given variables in each supercell.
}
\description{
Creates supercells by bringing the idea of superpixels into spatial analysis and grouping raster cells into coherent regions.
It uses a modified version of the SLIC algorithm by Achanta et al. (2012), allowing specification of a distance function.
}
\details{
If you want to use additional arguments for the averaging function (\code{avg_fun}), you can create a custom function. For example, if you want to calculate the mean by removing missing values, you can use the following code: \code{my_mean = function(x) mean(x, na.rm = TRUE)} and then provide \code{avg_fun = my_mean.}
For raster IDs or point centers outputs, see \code{\link[=sc_slic_raster]{sc_slic_raster()}} and
\code{\link[=sc_slic_points]{sc_slic_points()}}. For evaluation and diagnostics, see
\code{\link[=sc_metrics_pixels]{sc_metrics_pixels()}}, \code{\link[=sc_metrics_supercells]{sc_metrics_supercells()}}, and \code{\link[=sc_metrics_global]{sc_metrics_global()}}.
}
\examples{
library(supercells)
# One variable

vol = terra::rast(system.file("raster/volcano.tif", package = "supercells"))
vol_slic1 = supercells(vol, k = 50, compactness = 1)
terra::plot(vol)
plot(sf::st_geometry(vol_slic1), add = TRUE, lwd = 0.2)

# RGB variables
# ortho = terra::rast(system.file("raster/ortho.tif", package = "supercells"))
# ortho_slic1 = supercells(ortho, k = 1000, compactness = 10, transform = "to_LAB")
# terra::plot(ortho)
# plot(sf::st_geometry(ortho_slic1), add = TRUE)
#
# ### RGB variables - colored output
#
# rgb_to_hex = function(x){
#   apply(t(x), 2, function(x) rgb(x[1], x[2], x[3], maxColorValue = 255))
# }
# avg_colors = rgb_to_hex(sf::st_drop_geometry(ortho_slic1[4:6]))
#
# terra::plot(ortho)
# plot(sf::st_geometry(ortho_slic1), add = TRUE, col = avg_colors)
}
\references{
Achanta, R., Shaji, A., Smith, K., Lucchi, A., Fua, P., & Süsstrunk, S. (2012). SLIC Superpixels Compared to State-of-the-Art Superpixel Methods. IEEE Transactions on Pattern Analysis and Machine Intelligence, 34(11), 2274–2282. https://doi.org/10.1109/tpami.2012.120

Nowosad, J. Motif: an open-source R tool for pattern-based spatial analysis. Landscape Ecol (2021). https://doi.org/10.1007/s10980-020-01135-0
}
\seealso{
\code{\link[=sc_slic]{sc_slic()}}
}
