% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermotifs.R
\name{supermotifs}
\alias{supermotifs}
\title{Creates supermotifs}
\usage{
supermotifs(
  x,
  k,
  compactness,
  dist_fun = "jensen_shannon",
  clean = TRUE,
  iter = 10,
  step,
  minarea = 0
)
}
\arguments{
\item{x}{An object of class lsp (output of \code{motif::lsp_signature()} from the motif package)}

\item{k}{The number of supercells desired by the user (the output number can be slightly different!).
You can use either \code{k} or \code{step}.}

\item{compactness}{A compactness value. Larger values cause clusters to be more compact/even (square).
A compactness value depends on the range of input cell values and selected distance measure.}

\item{dist_fun}{A distance function. Currently implemented distance functions are "euclidean" and "jensen_shannon". Default: "jensen_shannon"}

\item{clean}{Should connectivity of the supercells be enforced?}

\item{iter}{The number of iterations performed to create the output.}

\item{step}{The distance (number of cells) between initial supercells' centers. You can use either \code{k} or \code{step}.}

\item{minarea}{Specifies the minimal size of a supercell (in cells). Only works when \code{clean = TRUE}.
By default, when \code{clean = TRUE}, average area (A) is calculated based on the total number of cells divided by a number of superpixels.
Next, the minimal size of a supercell equals to A/(2^2) (A is being right shifted)}
}
\value{
An sf object with several columns: (1) supercells - an id of each supercell, (2) y and x coordinates, (3) one or more columns with average values of given variables in each supercell
}
\description{
Creates supermotifs based on the lsp objects from the motif package. It uses a modified version of the SLIC Superpixel algorithm by Achanta et al. (2012), allowing specification of a distance function.
}
\examples{
if (requireNamespace("motif", quietly = TRUE) && requireNamespace("stars", quietly = TRUE)) {
  library(motif)
  library(stars)
  landcover = read_stars(system.file("raster/landcover2015.tif", package = "motif"))
  coma_output2 = lsp_signature(landcover, type = "cove", window = 10,
                               normalization = "pdf", ordered = FALSE)
  slic2 = supermotifs(coma_output2, k = 2000, compactness = 0.1, dist_fun = "jensen_shannon")

  plot(landcover, reset = FALSE)
  plot(st_geometry(slic2), add = TRUE)

}
}
\references{
Achanta, R., Shaji, A., Smith, K., Lucchi, A., Fua, P., & Süsstrunk, S. (2012). SLIC Superpixels Compared to State-of-the-Art Superpixel Methods. IEEE Transactions on Pattern Analysis and Machine Intelligence, 34(11), 2274–2282. https://doi.org/10.1109/tpami.2012.120
}
