% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_adaptive.R
\name{use_adaptive}
\alias{use_adaptive}
\title{Use adaptive compactness mode}
\usage{
use_adaptive(method = "local_max")
}
\arguments{
\item{method}{Adaptive compactness method. Currently only \code{"local_max"} is supported
(ASLIC-style).}
}
\value{
An adaptive compactness mode object for \code{compactness} arguments.
}
\description{
Creates a compactness mode object for adaptive compactness.
The \code{"local_max"} method corresponds to ASLIC-style local scaling,
where compactness is adapted using local maximum value distances.
}
\examples{
use_adaptive()

}
