% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_meters.R
\name{use_meters}
\alias{use_meters}
\title{Mark step values as meters}
\usage{
use_meters(x)
}
\arguments{
\item{x}{A single positive numeric value.}
}
\value{
A \link[units:units]{units::units} object in meters (\code{m}).
}
\description{
Creates a units value in meters for use in \code{step} arguments.
Use plain numerics for cell units, and \code{use_meters()} for map-distance steps.
}
\examples{
use_meters(100)

}
