% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-global.R
\name{sc_metrics_global}
\alias{sc_metrics_global}
\title{Global supercells metrics}
\usage{
sc_metrics_global(x, raster, dist_fun = "euclidean", compactness, step)
}
\arguments{
\item{x}{An sf object returned by \code{\link[=supercells]{supercells()}}.}

\item{raster}{The input SpatRaster used to create \code{x}.}

\item{dist_fun}{A distance function name or function, as in \code{\link[=supercells]{supercells()}}.}

\item{compactness}{A compactness value used for the supercells.}

\item{step}{A step value used for the supercells.}
}
\value{
A data.frame with a single row of global metrics.
}
\description{
Computes global distance diagnostics for supercells.
}
