% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercells-package.R
\docType{package}
\name{supercells-package}
\alias{supercells-package}
\title{supercells: Superpixels of Spatial Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates superpixels based on input spatial data. This package works on spatial data with one variable (e.g., continuous raster), many variables (e.g., RGB rasters), and spatial patterns (e.g., areas in categorical rasters). It is based on the SLIC algorithm (Achanta et al. (2012) \doi{10.1109/TPAMI.2012.120}), and readapts it to work with arbitrary dissimilarity measures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jakubnowosad.com/supercells/}
  \item Report bugs at \url{https://github.com/Nowosad/supercells/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

Other contributors:
\itemize{
  \item Pascal Mettes (Author of the initial C++ implementation of the SLIC Superpixel algorithm for image data) [contributor]
}

}
\keyword{internal}
