% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercells.R
\name{supercells}
\alias{supercells}
\title{Creates supercells}
\usage{
supercells(x, step, nc, dist_fun = "euclidean", clean = TRUE, iter = 10)
}
\arguments{
\item{x}{a}

\item{step}{a}

\item{nc}{a}

\item{dist_fun}{a}

\item{clean}{a}

\item{iter}{a}
}
\value{
a
}
\description{
Creates supercells
}
\examples{
#a
}
